@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: '>=3.3.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,21 +32,21 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  get: ^4.6.5
-  get_storage: ^2.0.3
-  carousel_slider: ^4.1.1
-  flutter_social_button: ^1.1.4
-  firebase_auth: ^3.5.0
-  firebase_core: ^1.20.0
-  cloud_firestore: ^3.1.7
-  google_sign_in: ^5.1.1
-  http: ^0.13.4
-  dotted_border: ^2.0.0+2
-  date_picker_timeline: ^1.2.3
-  lottie: ^2.0.0
-  palette_generator: ^0.3.3+2
-  youtube_player_flutter: ^8.1.1
+  cupertino_icons: ^1.0.6
+  get: ^4.6.6
+  get_storage: ^2.1.1
+  carousel_slider: ^4.2.1
+  flutter_social_button: ^1.1.4+1
+  firebase_auth: ^4.19.6
+  firebase_core: ^2.31.1
+  cloud_firestore: ^4.17.4
+  google_sign_in: ^6.2.1
+  http: ^1.2.1
+  dotted_border: ^2.1.0
+  lottie: ^3.1.2
+  palette_generator: ^0.3.3+3
+  youtube_player_flutter: ^9.0.1
+  date_picker_timeline: ^1.2.6
 
 dev_dependencies:
   flutter_test:
@@ -57,7 +57,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
